requirements_global.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
decord==0.6.0
distro==1.9.0
Django==5.1.3
django-appconf==1.0.6
dnspython==2.7.0
docstring_parser==0.16
easyocr==1.7.2
ecdsa==0.19.0
faiss-cpu==1.9.0.post1
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gradio==5.5.0
gradio_client==1.4.2
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imagekitio==4.1.0
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.6.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.3
langchain-google-genai==2.0.4
langchain-mongodb==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.139
lazy_loader==0.4
libclang==18.1.1
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
ocr_service==0.8
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pilkit==3.0
Pillow==10.1.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
pypdf==5.1.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruff==0.7.3
safehttpx==0.1.1
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.3.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.41.2
sympy==1.13.1
tabula==1.0.5
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
transformers==4.40.0
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.31.0
watchfiles==0.24.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.17.1
